requirements.txt
#
# This file is autogenerated by pip-compile with python 3.11
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
dj-database-url==1.0.0
    # via -r requirements.in
django==4.1.3
    # via
    #   -r requirements.in
    #   dj-database-url
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
psycopg2-binary==2.9.5
    # via -r requirements.in
pyjwt==2.6.0
    # via djangorestframework-simplejwt
python-decouple==3.6
    # via -r requirements.in
pytz==2022.6
    # via djangorestframework
sqlparse==0.4.3
    # via django

# The following packages are considered to be unsafe in a requirements file:
# setuptools
